Develop

+++ src/main/java/br/com/juwer/algafoodapi/infrastructure/repository/CidadeRepositoryImpl.java
@@ -5,7 +5,9 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import br.com.juwer.algafoodapi.domain.model.Cidade;
 import br.com.juwer.algafoodapi.domain.repository.CidadeRepository;
@@ -29,14 +31,20 @@ public class CidadeRepositoryImpl implements CidadeRepository {
     return entityManager.find(Cidade.class, id);
   }
 
+  @Transactional
   @Override
   public Cidade salvar(Cidade cidade) {
     return entityManager.merge(cidade);
   }
 
+  @Transactional
   @Override
-  public void remover(Cidade cidade) {
-    cidade = buscar(cidade.getId());
+  public void remover(Long cidadeId) {
+    Cidade cidade = buscar(cidadeId);
+
+    if(cidade == null){
+      throw new EmptyResultDataAccessException(null, 1);
+    }
     entityManager.remove(cidade);
   }
 }
